+++ pkg/injector/sidecar_container_test.go
@@ -86,229 +86,3 @@ func TestParseEnvString(t *testing.T) {
 		})
 	}
 }
-
-func TestParseVolumeMountsString(t *testing.T) {
-	testCases := []struct {
-		testName     string
-		mountStr     string
-		readOnly     bool
-		expMountsLen int
-		expMounts    []corev1.VolumeMount
-	}{
-		{
-			testName:     "empty volume mount string.",
-			mountStr:     "",
-			readOnly:     false,
-			expMountsLen: 0,
-			expMounts:    []corev1.VolumeMount{},
-		},
-		{
-			testName:     "valid volume mount string with readonly false.",
-			mountStr:     "my-mount:/tmp/mount1,another-mount:/home/user/mount2, mount3:/root/mount3",
-			readOnly:     false,
-			expMountsLen: 3,
-			expMounts: []corev1.VolumeMount{
-				{
-					Name:      "my-mount",
-					MountPath: "/tmp/mount1",
-				},
-				{
-					Name:      "another-mount",
-					MountPath: "/home/user/mount2",
-				},
-				{
-					Name:      "mount3",
-					MountPath: "/root/mount3",
-				},
-			},
-		},
-		{
-			testName:     "valid volume mount string with readonly true.",
-			mountStr:     " my-mount:/tmp/mount1,mount2:/root/mount2 ",
-			readOnly:     true,
-			expMountsLen: 2,
-			expMounts: []corev1.VolumeMount{
-				{
-					Name:      "my-mount",
-					MountPath: "/tmp/mount1",
-					ReadOnly:  true,
-				},
-				{
-					Name:      "mount2",
-					MountPath: "/root/mount2",
-					ReadOnly:  true,
-				},
-			},
-		},
-		{
-			testName:     "volume mount string with invalid mounts",
-			mountStr:     "my-mount:/tmp/mount1:rw,mount2:/root/mount2,mount3",
-			readOnly:     false,
-			expMountsLen: 1,
-			expMounts: []corev1.VolumeMount{
-				{
-					Name:      "mount2",
-					MountPath: "/root/mount2",
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.testName, func(t *testing.T) {
-			mounts := parseVolumeMountsString(tc.mountStr, tc.readOnly)
-			assert.Equal(t, tc.expMountsLen, len(mounts))
-			assert.Equal(t, tc.expMounts, mounts)
-		})
-	}
-}
-
-func TestPodContainsVolume(t *testing.T) {
-	testCases := []struct {
-		testName   string
-		podVolumes []corev1.Volume
-		volumeName string
-		expect     bool
-	}{
-		{
-			"pod with no volumes",
-			[]corev1.Volume{},
-			"volume1",
-			false,
-		},
-		{
-			"pod does not contain volume",
-			[]corev1.Volume{
-				{Name: "volume"},
-			},
-			"volume1",
-			false,
-		},
-		{
-			"pod contains volume",
-			[]corev1.Volume{
-				{Name: "volume1"},
-				{Name: "volume2"},
-			},
-			"volume2",
-			true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			pod := &corev1.Pod{
-				Spec: corev1.PodSpec{
-					Volumes: tc.podVolumes,
-				},
-			}
-			assert.Equal(t, tc.expect, podContainsVolume(pod, tc.volumeName))
-		})
-	}
-}
-
-func TestGetVolumeMounts(t *testing.T) {
-	testCases := []struct {
-		testName                  string
-		volumeReadOnlyAnnotation  string
-		volumeReadWriteAnnotation string
-		podVolumeMountNames       []string
-		expVolumeMounts           []corev1.VolumeMount
-	}{
-		{
-			"no annotations",
-			"",
-			"",
-			[]string{"mount1", "mount2"},
-			[]corev1.VolumeMount{},
-		},
-		{
-			"annotations with volumes present in pod",
-			"mount1:/tmp/mount1,mount2:/tmp/mount2",
-			"mount3:/tmp/mount3,mount4:/tmp/mount4",
-			[]string{"mount1", "mount2", "mount3", "mount4"},
-			[]corev1.VolumeMount{
-				{Name: "mount1", MountPath: "/tmp/mount1", ReadOnly: true},
-				{Name: "mount2", MountPath: "/tmp/mount2", ReadOnly: true},
-				{Name: "mount3", MountPath: "/tmp/mount3", ReadOnly: false},
-				{Name: "mount4", MountPath: "/tmp/mount4", ReadOnly: false},
-			},
-		},
-		{
-			"annotations with volumes not present in pod",
-			"mount1:/tmp/mount1,mount2:/tmp/mount2",
-			"mount3:/tmp/mount3,mount4:/tmp/mount4",
-			[]string{"mount1", "mount2", "mount4"},
-			[]corev1.VolumeMount{
-				{Name: "mount1", MountPath: "/tmp/mount1", ReadOnly: true},
-				{Name: "mount2", MountPath: "/tmp/mount2", ReadOnly: true},
-				{Name: "mount4", MountPath: "/tmp/mount4", ReadOnly: false},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			pod := &corev1.Pod{
-				Spec: corev1.PodSpec{
-					Volumes: []corev1.Volume{},
-				},
-			}
-			for _, volumeName := range tc.podVolumeMountNames {
-				pod.Spec.Volumes = append(pod.Spec.Volumes, corev1.Volume{Name: volumeName})
-			}
-
-			c := NewSidecarConfig(pod)
-			c.VolumeMounts = tc.volumeReadOnlyAnnotation
-			c.VolumeMountsRW = tc.volumeReadWriteAnnotation
-
-			volumeMounts := c.GetVolumeMounts()
-			assert.Equal(t, tc.expVolumeMounts, volumeMounts)
-		})
-	}
-}
-
-func TestGetUnixDomainSocketVolume(t *testing.T) {
-	testCases := []struct {
-		testName        string
-		udsPath         string
-		originalVolumes []corev1.Volume
-		exportMount     *corev1.VolumeMount
-	}{
-		{
-			"empty value",
-			"",
-			nil,
-			nil,
-		},
-		{
-			"add volume",
-			"/tmp",
-			nil,
-			&corev1.VolumeMount{Name: UnixDomainSocketVolume, MountPath: "/tmp"},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			pod := &corev1.Pod{
-				Spec: corev1.PodSpec{
-					Volumes: tc.originalVolumes,
-				},
-			}
-
-			c := NewSidecarConfig(pod)
-			c.UnixDomainSocketPath = tc.udsPath
-
-			socketMount := c.GetUnixDomainSocketVolumeMount()
-
-			if tc.exportMount == nil {
-				assert.Equal(t, tc.exportMount, socketMount)
-			} else {
-				assert.Equal(t, tc.exportMount.Name, socketMount.Name)
-				assert.Equal(t, tc.exportMount.MountPath, socketMount.MountPath)
-			}
-		})
-	}
-}
