+++ pkg/injector/sidecar_patcher.go
@@ -1,6 +1,8 @@
 package injector
 
 import (
+	"strconv"
+
 	jsonpatch "github.com/evanphx/json-patch/v5"
 	corev1 "k8s.io/api/core/v1"
 
@@ -21,51 +23,32 @@ func (c *SidecarConfig) GetPatch() (patchOps jsonpatch.Patch, err error) {
 		return nil, err
 	}
 
-	// Get all volume mounts
-	volumeMounts := sidecar.GetVolumeMounts(pod)
-	socketVolumeMount := sidecar.GetUnixDomainSocketVolumeMount(&pod)
+	// Get volume mounts and add the UDS volume mount if needed
+	volumeMounts := c.GetVolumeMounts()
+	socketVolumeMount := c.GetUnixDomainSocketVolumeMount()
 	if socketVolumeMount != nil {
 		volumeMounts = append(volumeMounts, *socketVolumeMount)
 	}
-	volumeMounts = append(volumeMounts, corev1.VolumeMount{
-		Name:      sidecar.TokenVolumeName,
-		MountPath: sidecar.TokenVolumeKubernetesMountPath,
-		ReadOnly:  true,
-	})
 
 	// Pluggable components
 	appContainers, componentContainers, injectedComponentContainers, err := i.splitContainers(pod)
 	if err != nil {
 		return nil, err
 	}
-	componentPatchOps, componentsSocketVolumeMount := components.PatchOps(componentContainers, injectedComponentContainers, &pod)
+	componentPatchOps, componentsSocketVolumeMount := components.PatchOps(componentContainers, injectedComponentContainers, c.pod)
 
 	// Projected volume with the token
-	tokenVolume := sidecar.GetTokenVolume()
+	tokenVolume := c.GetTokenVolume()
+	volumeMounts = append(volumeMounts, corev1.VolumeMount{
+		Name:      TokenVolumeName,
+		MountPath: TokenVolumeKubernetesMountPath,
+		ReadOnly:  true,
+	})
 
 	// Get the sidecar container
-	sidecarContainer, err := sidecar.GetSidecarContainer(sidecar.ContainerConfig{
-		AppID:                        appID,
-		Annotations:                  an,
-		CertChain:                    certChain,
-		CertKey:                      certKey,
-		ControlPlaneAddress:          apiSvcAddress,
-		DaprSidecarImage:             image,
-		Identity:                     req.Namespace + ":" + pod.Spec.ServiceAccountName,
-		IgnoreEntrypointTolerations:  i.config.GetIgnoreEntrypointTolerations(),
-		ImagePullPolicy:              i.config.GetPullPolicy(),
-		MTLSEnabled:                  mTLSEnabled(daprClient),
-		Namespace:                    req.Namespace,
-		PlacementServiceAddress:      placementAddress,
-		SentryAddress:                sentryAddress,
-		Tolerations:                  pod.Spec.Tolerations,
-		TrustAnchors:                 trustAnchors,
-		VolumeMounts:                 volumeMounts,
+	sidecarContainer, err := c.GetSidecarContainer(getSidecarContainerOpts{
 		ComponentsSocketsVolumeMount: componentsSocketVolumeMount,
-		SkipPlacement:                i.config.GetSkipPlacement(),
-		RunAsNonRoot:                 i.config.GetRunAsNonRoot(),
-		ReadOnlyRootFilesystem:       i.config.GetReadOnlyRootFilesystem(),
-		SidecarDropALLCapabilities:   i.config.GetDropCapabilities(),
+		VolumeMounts:                 volumeMounts,
 	})
 	if err != nil {
 		return nil, err
@@ -73,7 +56,7 @@ func (c *SidecarConfig) GetPatch() (patchOps jsonpatch.Patch, err error) {
 
 	// Create the list of patch operations
 	patchOps = jsonpatch.Patch{}
-	if len(pod.Spec.Containers) == 0 {
+	if len(c.pod.Spec.Containers) == 0 {
 		// Set to empty to support add operations individually
 		patchOps = append(patchOps,
 			patcher.NewPatchOperation("add", patcher.PatchPathContainers, []corev1.Container{}),
@@ -82,19 +65,19 @@ func (c *SidecarConfig) GetPatch() (patchOps jsonpatch.Patch, err error) {
 
 	patchOps = append(patchOps,
 		patcher.NewPatchOperation("add", patcher.PatchPathContainers+"/-", sidecarContainer),
-		sidecar.AddDaprSidecarInjectedLabel(pod.Labels),
-		sidecar.AddDaprSidecarAppIDLabel(c.GetAppID(), c.pod.Labels),
-		sidecar.AddDaprSidecarMetricsEnabledLabel(c.EnableMetrics, c.pod.Labels),
+		AddDaprSidecarInjectedLabel(c.pod.Labels),
+		AddDaprSidecarAppIDLabel(c.GetAppID(), c.pod.Labels),
+		AddDaprSidecarMetricsEnabledLabel(c.EnableMetrics, c.pod.Labels),
 	)
 
 	patchOps = append(patchOps,
-		sidecar.AddDaprEnvVarsToContainers(appContainers, getAppProtocol(an))...,
+		AddDaprEnvVarsToContainers(appContainers, getAppProtocol(an))...,
 	)
 	patchOps = append(patchOps,
-		sidecar.AddSocketVolumeMountToContainers(appContainers, socketVolumeMount)...,
+		AddSocketVolumeMountToContainers(appContainers, socketVolumeMount)...,
 	)
-	volumePatchOps := sidecar.GetVolumesPatchOperations(
-		pod.Spec.Volumes,
+	volumePatchOps := GetVolumesPatchOperations(
+		c.pod.Spec.Volumes,
 		[]corev1.Volume{tokenVolume},
 		patcher.PatchPathVolumes,
 	)
@@ -112,4 +95,62 @@ func (c *SidecarConfig) podContainsSidecarContainer() bool {
 		}
 	}
 	return false
+}
+
+// AddDaprEnvVarsToContainers adds Dapr environment variables to all the containers in any Dapr-enabled pod.
+// The containers can be injected or user-defined.
+func (c *SidecarConfig) AddDaprEnvVarsToContainers(containers map[int]corev1.Container, appProtocol string) jsonpatch.Patch {
+	envPatchOps := make(jsonpatch.Patch, 0, len(containers)*2)
+	envVars := []corev1.EnvVar{
+		{
+			Name:  UserContainerDaprHTTPPortName,
+			Value: strconv.FormatInt(int64(c.SidecarHTTPPort), 10),
+		},
+		{
+			Name:  UserContainerDaprGRPCPortName,
+			Value: strconv.FormatInt(int64(c.SidecarAPIGRPCPort), 10),
+		},
+	}
+	if appProtocol != "" {
+		envVars = append(envVars, corev1.EnvVar{
+			Name:  UserContainerAppProtocolName,
+			Value: appProtocol,
+		})
+	}
+	for i, container := range containers {
+		patchOps := patcher.GetEnvPatchOperations(container.Env, envVars, i)
+		envPatchOps = append(envPatchOps, patchOps...)
+	}
+	return envPatchOps
+}
+
+// AddDaprSidecarInjectedLabel adds Dapr label to patch pod so list of patched pods can be retrieved more efficiently
+func AddDaprSidecarInjectedLabel(labels map[string]string) jsonpatch.Operation {
+	if len(labels) == 0 { // empty labels
+		return patcher.NewPatchOperation("add", PatchPathLabels, map[string]string{
+			SidecarInjectedLabel: "true",
+		})
+	}
+
+	return patcher.NewPatchOperation("add", PatchPathLabels+"/dapr.io~1sidecar-injected", "true")
+}
+
+// AddDaprSidecarAppIDLabel adds Dapr app-id label which can be handy for metric labels
+func AddDaprSidecarAppIDLabel(appID string, labels map[string]string) jsonpatch.Operation {
+	if len(labels) == 0 { // empty labels
+		return patcher.NewPatchOperation("add", PatchPathLabels, map[string]string{
+			SidecarAppIDLabel: appID,
+		})
+	}
+	return patcher.NewPatchOperation("add", PatchPathLabels+"/dapr.io~1app-id", appID)
+}
+
+// AddDaprSidecarMetricsEnabledLabel adds Dapr metrics-enabled label which can be handy for scraping metrics
+func AddDaprSidecarMetricsEnabledLabel(metricsEnabled bool, labels map[string]string) jsonpatch.Operation {
+	if len(labels) == 0 { // empty labels
+		return patcher.NewPatchOperation("add", PatchPathLabels, map[string]string{
+			SidecarMetricsEnabledLabel: strconv.FormatBool(metricsEnabled),
+		})
+	}
+	return patcher.NewPatchOperation("add", PatchPathLabels+"/dapr.io~1metrics-enabled", strconv.FormatBool(metricsEnabled))
 }
