+++ pkg/injector/sidecar_patcher_test.go
@@ -1,27 +1,12 @@
-/*
-Copyright 2022 The Dapr Authors
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package sidecar
+package injector
 
 import (
 	"encoding/json"
-	"strconv"
 	"testing"
 
 	jsonpatch "github.com/evanphx/json-patch/v5"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	coreV1 "k8s.io/api/core/v1"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/dapr/dapr/pkg/injector/patcher"
@@ -30,35 +15,35 @@ import (
 func TestAddDaprEnvVarsToContainers(t *testing.T) {
 	testCases := []struct {
 		testName      string
-		mockContainer coreV1.Container
+		mockContainer corev1.Container
 		appProtocol   string
 		expOpsLen     int
 		expOps        jsonpatch.Patch
 	}{
 		{
 			testName: "empty environment vars",
-			mockContainer: coreV1.Container{
+			mockContainer: corev1.Container{
 				Name: "MockContainer",
 			},
 			expOpsLen: 1,
 			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/env", []coreV1.EnvVar{
+				patcher.NewPatchOperation("add", "/spec/containers/0/env", []corev1.EnvVar{
 					{
 						Name:  UserContainerDaprHTTPPortName,
-						Value: strconv.Itoa(SidecarHTTPPort),
+						Value: "3500",
 					},
 					{
 						Name:  UserContainerDaprGRPCPortName,
-						Value: strconv.Itoa(SidecarAPIGRPCPort),
+						Value: "50001",
 					},
 				}),
 			},
 		},
 		{
 			testName: "existing env var",
-			mockContainer: coreV1.Container{
+			mockContainer: corev1.Container{
 				Name: "Mock Container",
-				Env: []coreV1.EnvVar{
+				Env: []corev1.EnvVar{
 					{
 						Name:  "TEST",
 						Value: "Existing value",
@@ -67,21 +52,21 @@ func TestAddDaprEnvVarsToContainers(t *testing.T) {
 			},
 			expOpsLen: 2,
 			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/env/-", coreV1.EnvVar{
+				patcher.NewPatchOperation("add", "/spec/containers/0/env/-", corev1.EnvVar{
 					Name:  UserContainerDaprHTTPPortName,
-					Value: strconv.Itoa(SidecarHTTPPort),
+					Value: "3500",
 				}),
-				patcher.NewPatchOperation("add", "/spec/containers/0/env/-", coreV1.EnvVar{
+				patcher.NewPatchOperation("add", "/spec/containers/0/env/-", corev1.EnvVar{
 					Name:  UserContainerDaprGRPCPortName,
-					Value: strconv.Itoa(SidecarAPIGRPCPort),
+					Value: "50001",
 				}),
 			},
 		},
 		{
 			testName: "existing conflicting env var",
-			mockContainer: coreV1.Container{
+			mockContainer: corev1.Container{
 				Name: "Mock Container",
-				Env: []coreV1.EnvVar{
+				Env: []corev1.EnvVar{
 					{
 						Name:  "TEST",
 						Value: "Existing value",
@@ -94,17 +79,17 @@ func TestAddDaprEnvVarsToContainers(t *testing.T) {
 			},
 			expOpsLen: 1,
 			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/env/-", coreV1.EnvVar{
+				patcher.NewPatchOperation("add", "/spec/containers/0/env/-", corev1.EnvVar{
 					Name:  UserContainerDaprHTTPPortName,
-					Value: strconv.Itoa(SidecarHTTPPort),
+					Value: "3500",
 				}),
 			},
 		},
 		{
 			testName: "multiple existing conflicting env vars",
-			mockContainer: coreV1.Container{
+			mockContainer: corev1.Container{
 				Name: "Mock Container",
-				Env: []coreV1.EnvVar{
+				Env: []corev1.EnvVar{
 					{
 						Name:  UserContainerDaprHTTPPortName,
 						Value: "3510",
@@ -120,20 +105,20 @@ func TestAddDaprEnvVarsToContainers(t *testing.T) {
 		},
 		{
 			testName: "with app protocol",
-			mockContainer: coreV1.Container{
+			mockContainer: corev1.Container{
 				Name: "MockContainer",
 			},
 			expOpsLen:   1,
 			appProtocol: "h2c",
 			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/env", []coreV1.EnvVar{
+				patcher.NewPatchOperation("add", "/spec/containers/0/env", []corev1.EnvVar{
 					{
 						Name:  UserContainerDaprHTTPPortName,
-						Value: strconv.Itoa(SidecarHTTPPort),
+						Value: "3500",
 					},
 					{
 						Name:  UserContainerDaprGRPCPortName,
-						Value: strconv.Itoa(SidecarAPIGRPCPort),
+						Value: "50001",
 					},
 					{
 						Name:  UserContainerAppProtocolName,
@@ -146,7 +131,8 @@ func TestAddDaprEnvVarsToContainers(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.testName, func(t *testing.T) {
-			patchEnv := AddDaprEnvVarsToContainers(map[int]coreV1.Container{0: tc.mockContainer}, tc.appProtocol)
+			c := NewSidecarConfig(&corev1.Pod{})
+			patchEnv := c.AddDaprEnvVarsToContainers(map[int]corev1.Container{0: tc.mockContainer}, tc.appProtocol)
 			assert.Equal(t, tc.expOpsLen, len(patchEnv))
 			assert.Equal(t, tc.expOps, patchEnv)
 		})
@@ -156,19 +142,19 @@ func TestAddDaprEnvVarsToContainers(t *testing.T) {
 func TestAddDaprAppIDLabel(t *testing.T) {
 	testCases := []struct {
 		testName  string
-		mockPod   coreV1.Pod
+		mockPod   corev1.Pod
 		expLabels map[string]string
 	}{
 		{
 			testName: "empty labels",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{},
 			},
 			expLabels: map[string]string{SidecarAppIDLabel: "my-app"},
 		},
 		{
 			testName: "with some previous labels",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Labels: map[string]string{"app": "my-app"},
 				},
@@ -177,7 +163,7 @@ func TestAddDaprAppIDLabel(t *testing.T) {
 		},
 		{
 			testName: "with dapr app-id label already present",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Labels: map[string]string{SidecarAppIDLabel: "my-app", "app": "my-app"},
 				},
@@ -192,7 +178,7 @@ func TestAddDaprAppIDLabel(t *testing.T) {
 			newPodJSON := patchObject(t, tc.mockPod, jsonpatch.Patch{
 				AddDaprSidecarAppIDLabel("my-app", tc.mockPod.Labels),
 			})
-			newPod := coreV1.Pod{}
+			newPod := corev1.Pod{}
 			assert.NoError(t, json.Unmarshal(newPodJSON, &newPod))
 			assert.Equal(t, tc.expLabels, newPod.Labels)
 		})
@@ -202,13 +188,13 @@ func TestAddDaprAppIDLabel(t *testing.T) {
 func TestAddDaprMetricsEnabledLabel(t *testing.T) {
 	testCases := []struct {
 		testName       string
-		mockPod        coreV1.Pod
+		mockPod        corev1.Pod
 		expLabels      map[string]string
 		metricsEnabled bool
 	}{
 		{
 			testName: "metrics annotation not present, fallback to default",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{},
 			},
 			expLabels:      map[string]string{SidecarMetricsEnabledLabel: "false"},
@@ -216,7 +202,7 @@ func TestAddDaprMetricsEnabledLabel(t *testing.T) {
 		},
 		{
 			testName: "metrics annotation present and explicitly enabled, with existing labels",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Annotations: map[string]string{SidecarMetricsEnabledLabel: "true"},
 					Labels:      map[string]string{"app": "my-app"},
@@ -227,7 +213,7 @@ func TestAddDaprMetricsEnabledLabel(t *testing.T) {
 		},
 		{
 			testName: "metrics annotation present and explicitly disabled",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Annotations: map[string]string{SidecarMetricsEnabledLabel: "false"},
 				},
@@ -243,7 +229,7 @@ func TestAddDaprMetricsEnabledLabel(t *testing.T) {
 			newPodJSON := patchObject(t, tc.mockPod, jsonpatch.Patch{
 				AddDaprSidecarMetricsEnabledLabel(tc.metricsEnabled, tc.mockPod.Labels),
 			})
-			newPod := coreV1.Pod{}
+			newPod := corev1.Pod{}
 			assert.NoError(t, json.Unmarshal(newPodJSON, &newPod))
 			assert.Equal(t, tc.expLabels, newPod.Labels)
 		})
@@ -253,19 +239,19 @@ func TestAddDaprMetricsEnabledLabel(t *testing.T) {
 func TestAddDaprInjectedLabel(t *testing.T) {
 	testCases := []struct {
 		testName  string
-		mockPod   coreV1.Pod
+		mockPod   corev1.Pod
 		expLabels map[string]string
 	}{
 		{
 			testName: "empty labels",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{},
 			},
 			expLabels: map[string]string{SidecarInjectedLabel: "true"},
 		},
 		{
 			testName: "with some previous labels",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Labels: map[string]string{"app": "my-app"},
 				},
@@ -274,7 +260,7 @@ func TestAddDaprInjectedLabel(t *testing.T) {
 		},
 		{
 			testName: "with dapr injected label already present",
-			mockPod: coreV1.Pod{
+			mockPod: corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Labels: map[string]string{SidecarInjectedLabel: "true", "app": "my-app"},
 				},
@@ -289,134 +275,9 @@ func TestAddDaprInjectedLabel(t *testing.T) {
 			newPodJSON := patchObject(t, tc.mockPod, jsonpatch.Patch{
 				AddDaprSidecarInjectedLabel(tc.mockPod.Labels)},
 			)
-			newPod := coreV1.Pod{}
+			newPod := corev1.Pod{}
 			assert.NoError(t, json.Unmarshal(newPodJSON, &newPod))
 			assert.Equal(t, tc.expLabels, newPod.Labels)
-		})
-	}
-}
-
-// patchObject executes a jsonpatch action against the object passed
-func patchObject(t *testing.T, origObj interface{}, patch jsonpatch.Patch) []byte {
-	podJSON, err := json.Marshal(origObj)
-	require.NoError(t, err)
-	newJSON, err := patch.Apply(podJSON)
-	require.NoError(t, err)
-	return newJSON
-}
-
-func TestAddSocketVolumeToContainers(t *testing.T) {
-	testCases := []struct {
-		testName      string
-		mockContainer coreV1.Container
-		socketMount   *coreV1.VolumeMount
-		expOpsLen     int
-		expOps        jsonpatch.Patch
-	}{
-		{
-			testName: "empty var, empty volume",
-			mockContainer: coreV1.Container{
-				Name: "MockContainer",
-			},
-			socketMount: nil,
-			expOpsLen:   0,
-			expOps:      jsonpatch.Patch{},
-		},
-		{
-			testName: "existing var, empty volume",
-			mockContainer: coreV1.Container{
-				Name: "MockContainer",
-			},
-			socketMount: &coreV1.VolumeMount{
-				Name:      UnixDomainSocketVolume,
-				MountPath: "/tmp",
-			},
-			expOpsLen: 1,
-			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/volumeMounts", []coreV1.VolumeMount{{
-					Name:      UnixDomainSocketVolume,
-					MountPath: "/tmp",
-				}}),
-			},
-		},
-		{
-			testName: "existing var, existing volume",
-			mockContainer: coreV1.Container{
-				Name: "MockContainer",
-				VolumeMounts: []coreV1.VolumeMount{
-					{Name: "mock1"},
-				},
-			},
-			socketMount: &coreV1.VolumeMount{
-				Name:      UnixDomainSocketVolume,
-				MountPath: "/tmp",
-			},
-			expOpsLen: 1,
-			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/volumeMounts/-", coreV1.VolumeMount{
-					Name:      UnixDomainSocketVolume,
-					MountPath: "/tmp",
-				}),
-			},
-		},
-		{
-			testName: "existing var, multiple existing volumes",
-			mockContainer: coreV1.Container{
-				Name: "MockContainer",
-				VolumeMounts: []coreV1.VolumeMount{
-					{Name: "mock1"},
-					{Name: "mock2"},
-				},
-			},
-			socketMount: &coreV1.VolumeMount{
-				Name:      UnixDomainSocketVolume,
-				MountPath: "/tmp",
-			},
-			expOpsLen: 1,
-			expOps: jsonpatch.Patch{
-				patcher.NewPatchOperation("add", "/spec/containers/0/volumeMounts/-", coreV1.VolumeMount{
-					Name:      UnixDomainSocketVolume,
-					MountPath: "/tmp",
-				}),
-			},
-		},
-		{
-			testName: "existing var, conflict volume name",
-			mockContainer: coreV1.Container{
-				Name: "MockContainer",
-				VolumeMounts: []coreV1.VolumeMount{
-					{Name: UnixDomainSocketVolume},
-				},
-			},
-			socketMount: &coreV1.VolumeMount{
-				Name:      UnixDomainSocketVolume,
-				MountPath: "/tmp",
-			},
-			expOpsLen: 0,
-			expOps:    jsonpatch.Patch{},
-		},
-		{
-			testName: "existing var, conflict volume mount path",
-			mockContainer: coreV1.Container{
-				Name: "MockContainer",
-				VolumeMounts: []coreV1.VolumeMount{
-					{MountPath: "/tmp"},
-				},
-			},
-			socketMount: &coreV1.VolumeMount{
-				Name:      UnixDomainSocketVolume,
-				MountPath: "/tmp",
-			},
-			expOpsLen: 0,
-			expOps:    jsonpatch.Patch{},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			patchEnv := AddSocketVolumeMountToContainers(map[int]coreV1.Container{0: tc.mockContainer}, tc.socketMount)
-			assert.Equal(t, tc.expOpsLen, len(patchEnv))
-			assert.Equal(t, tc.expOps, patchEnv)
 		})
 	}
 }
