+++ pkg/injector/sidecar_test.go
@@ -14,11 +14,15 @@ limitations under the License.
 package injector
 
 import (
+	"encoding/json"
 	"testing"
 
-	"github.com/dapr/dapr/pkg/injector/annotations"
+	jsonpatch "github.com/evanphx/json-patch/v5"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
+
+	"github.com/dapr/dapr/pkg/injector/annotations"
 )
 
 func TestSidecarConfigInit(t *testing.T) {
@@ -62,4 +66,13 @@ func TestSidecarConfigSetFromAnnotations(t *testing.T) {
 	// Nullable properties
 	_ = assert.NotNil(t, c.EnableAPILogging) &&
 		assert.False(t, *c.EnableAPILogging)
+}
+
+// patchObject executes a jsonpatch action against the object passed
+func patchObject(t *testing.T, origObj interface{}, patch jsonpatch.Patch) []byte {
+	podJSON, err := json.Marshal(origObj)
+	require.NoError(t, err)
+	newJSON, err := patch.Apply(podJSON)
+	require.NoError(t, err)
+	return newJSON
 }
